requirements.txt
streamlit
google-generativeai
python-dotenv

torch
diffusers
transformers